# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application Programming Interface (API) for Pasiap Apps",
        "description": "L5 Swagger OpenAPI documentation for Pasiap Apps",
        "contact": {
            "email": "burhanburdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/pasiap/public/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/profiles": {
            "get": {
                "tags": [
                    "Profiles"
                ],
                "description": "Get data profile account",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profiles"
                ],
                "description": "Update data profile account",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "sid": {
                                        "type": "string"
                                    },
                                    "birth_place": {
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "type": "string"
                                    },
                                    "sex": {
                                        "type": "string"
                                    },
                                    "religion": {
                                        "type": "string"
                                    },
                                    "marital_status": {
                                        "type": "string"
                                    },
                                    "identity_card_photo": {
                                        "type": "file"
                                    },
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tokens": {
            "post": {
                "tags": [
                    "Profiles"
                ],
                "description": "Store token firebase into profile account",
                "operationId": "storeTokenFcm",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}